fix(domain): align double_spend_proof version parameter with sibling message classes
`double_spend_proof::to_data(size_t)` and `serialized_size(size_t)`
were the only two call sites in the message family that used
`size_t` for the protocol version. Every sibling (`merkle_block`,
`compact_block`, `get_blocks`, `get_headers`,
`prefilled_transaction`, `double_spend_proof::spender`, ...) takes
`uint32_t`, and so does `double_spend_proof::from_data` — the
version field itself is declared `uint32_t version = 0;` inside
`spender`.

The inconsistency is load-bearing downstream: the C-API generator
pipes `size_t` through as `kth_size_t` and `kth_size_t` maps to
`Py_ssize_t` on the py-native side, so the Python wrapper for DSP
ended up parsing a non-negative protocol version as a signed
Py_ssize_t with a non-negativity check that's structurally
unreachable for an unsigned type. Every other message binding
already parses it as `unsigned int` / `"I"` and casts to
`uint32_t`.

Change the four DSP entry points — `to_data(uint32_t) const`,
`to_data(uint32_t, data_sink&) const`, the templated
`to_data(uint32_t, W&)`, and `serialized_size(uint32_t) const` —
to take `uint32_t`, matching the sibling classes. Implementation
signatures in `double_spend_proof.cpp` follow the header.

Tests call `dsp.to_data(0)` / `dsp.serialized_size(0)` with the
integer literal `0`, which works unchanged under either type, so
the change is source-compatible for in-tree callers.

C-API regenerated through the generator pipeline: the three entry
points that previously carried `kth_size_t version` now carry
`uint32_t version`, matching the signature shape that
`merkle_block`, `compact_block`, `get_blocks`, `get_headers`,
`prefilled_transaction` and `double_spend_proof::spender`
already ship.

+++ src/domain/include/kth/domain/message/double_spend_proof.hpp
@@ -129,12 +129,12 @@ struct KD_API double_spend_proof {
     void set_spender2(spender const& x);
 
     [[nodiscard]]
-    data_chunk to_data(size_t /*version*/) const;
+    data_chunk to_data(uint32_t /*version*/) const;
 
-    void to_data(size_t /*version*/, data_sink& stream) const;
+    void to_data(uint32_t /*version*/, data_sink& stream) const;
 
     template <typename W>
-    void to_data(size_t /*version*/, W& sink) const {
+    void to_data(uint32_t /*version*/, W& sink) const {
         out_point_.to_data(sink);
         spender1_.to_data(sink);
         spender2_.to_data(sink);
@@ -149,7 +149,7 @@ struct KD_API double_spend_proof {
     void reset();
 
     [[nodiscard]]
-    size_t serialized_size(size_t /*version*/) const {
+    size_t serialized_size(uint32_t /*version*/) const {
         return out_point_.serialized_size() +
             spender1_.serialized_size() +
             spender2_.serialized_size();
